+++ src/actor/buffer.jl
@@ -24,7 +24,7 @@ show(getvalues(actor))
 
 See also: [`Actor`](@ref), [`buffer`](@ref)
 """
-struct BufferActor{T, R} <: Actor{R}
+struct BufferActor{T, R}
     values :: R
 end
 
@@ -37,9 +37,9 @@ end
 
 getvalues(actor::BufferActor) = actor.values
 
-next!(actor::BufferActor{T, R}, data::R) where {T, R} = copyto!(actor.values, data)
-error!(actor::BufferActor, err)                       = error(err)
-complete!(actor::BufferActor)                         = begin end
+on_next!(actor::BufferActor{T, R}, data::R) where {T, R} = copyto!(actor.values, data)
+on_error!(actor::BufferActor, err)                       = error(err)
+on_complete!(actor::BufferActor)                         = begin end
 
 """
     buffer(::Type{T}, size::Int) where T