+++ src/actor/logger.jl
@@ -13,7 +13,7 @@ The `LoggerActor` logs all `next!`/`error!`/`complete!` events that are sent fro
 
 See also: [`Actor`](@ref), [`logger`](@ref)
 """
-struct LoggerActor{O} <: Actor{Any}
+struct LoggerActor{O}
     name :: String
     io   :: O
 
@@ -24,17 +24,17 @@ Base.show(io::IO, actor::LoggerActor) = print(io, "LoggerActor(\"$(actor.name)\"
 
 # Remark: Here nothing for `io` is a workaround for https://github.com/JuliaDocs/Documenter.jl/issues/1245 where println(actor.io, ...) fails on doctest even if actor.io === stdout
 
-next!(actor::LoggerActor, data) = next!(actor.io, actor, data)
-error!(actor::LoggerActor, err) = error!(actor.io, actor, err)
-complete!(actor::LoggerActor)   = complete!(actor.io, actor)
+on_next!(actor::LoggerActor, data) = on_next!(actor.io, actor, data)
+on_error!(actor::LoggerActor, err) = on_error!(actor.io, actor, err)
+on_complete!(actor::LoggerActor)   = on_complete!(actor.io, actor)
 
-next!(::Nothing, actor::LoggerActor, data) = next!(stdout, actor, data)
-error!(::Nothing, actor::LoggerActor, err) = error!(stdout, actor, err)
-complete!(::Nothing, actor::LoggerActor)   = complete!(stdout, actor)
+on_next!(::Nothing, actor::LoggerActor, data) = on_next!(stdout, actor, data)
+on_error!(::Nothing, actor::LoggerActor, err) = on_error!(stdout, actor, err)
+on_complete!(::Nothing, actor::LoggerActor)   = on_complete!(stdout, actor)
 
-next!(io::IO, actor::LoggerActor, data) = println(io, "[$(actor.name)] Data: $data")
-error!(io::IO, actor::LoggerActor, err) = println(io, "[$(actor.name)] Error: $err")
-complete!(io::IO, actor::LoggerActor)   = println(io, "[$(actor.name)] Completed")
+on_next!(io::IO, actor::LoggerActor, data) = println(io, "[$(actor.name)] Data: $data")
+on_error!(io::IO, actor::LoggerActor, err) = println(io, "[$(actor.name)] Error: $err")
+on_complete!(io::IO, actor::LoggerActor)   = println(io, "[$(actor.name)] Completed")
 
 """
     logger([ io::IO ], name::String = "LogActor")