+++ src/operators/filter_type.jl
@@ -3,21 +3,21 @@ export filter_type
 import Base: show
 
 """
-    filter_type(type::Type{ T }) where { T }
+    filter_type(type::Type{R}) where R
 
 Creates a `filter_type` operator, which filters items of the source Observable by emitting only
-those that match a specified `T` type with a `<:` operator. This operator is a more efficient version of
-`filter(v -> v <: T) |> map(T, v -> v)` operators chain.
+those that match a specified `R` type with a `<:` operator. This operator is a more efficient version of
+`filter(v -> v <: R) |> map(OpType(R), v -> v)` operators chain.
 
 # Producing
 
-Stream of type `<: Subscribable{ T }` where `T` refers to `type` argument
+Stream of type `<: Subscribable{R}` where `R` refers to the `type` argument
 
 # Examples
 ```jldoctest
 using Rocket
 
-source = from(Real[ 1, 2.0, 3, 4.0, 5, 6.0 ])
+source = from_iterable(Real[ 1, 2.0, 3, 4.0, 5, 6.0 ])
 subscribe!(source |> filter_type(Int), logger())
 ;
 
@@ -29,34 +29,38 @@ subscribe!(source |> filter_type(Int), logger())
 [LogActor] Completed
 ```
 
-See also: [`AbstractOperator`](@ref), [`InferableOperator`](@ref), [`ProxyObservable`](@ref), [`logger`](@ref)
+See also: [`Operator`](@ref), [`from_iterable`](@ref), [`logger`](@ref)
 """
-filter_type(::Type{ T }) where { T }  = FilterTypeOperator{T}()
+filter_type(::Type{R}) where R  = FilterTypeOperator{R}()
 
-struct FilterTypeOperator{T} <: InferableOperator end
+struct FilterTypeOperator{R} <: Operator end
 
-on_call!(::Type{T}, ::Type{T}, operator::FilterTypeOperator{T}, source) where { T    }      = source
-on_call!(::Type{L}, ::Type{T}, operator::FilterTypeOperator{T}, source) where { L, T <: L } = proxy(T, source, FilterTypeProxy{L, T}())
-on_call!(::Type{L}, ::Type{T}, operator::FilterTypeOperator{T}, source) where { L, T }      = error("filter_type(T) operator usage error. Cannot filter $(T) elements from stream of $(L) elements.")
+on_call!(::Type{R}, ::Type{R}, operator::FilterTypeOperator{R}, source)    where { R    }         = source
+on_call!(::Type{L}, ::Type{R}, operator::FilterTypeOperator{R}, source::S) where { L, R <: L, S } = FilterTypeSubscribable{R, S}(source)
+on_call!(::Type{L}, ::Type{R}, operator::FilterTypeOperator{R}, source)    where { L, R }         = error("filter_type(T) operator usage error. Cannot filter $(R) elements from stream of $(L) elements.")
 
-operator_right(::FilterTypeOperator{T}, ::Type{T}) where { T }         = T
-operator_right(::FilterTypeOperator{T}, ::Type{L}) where { L, T <: L } = T
-operator_right(::FilterTypeOperator{T}, ::Type{L}) where { L, T }      = error("filter_type(T) operator creation error. Cannot filter $(T) elements from stream of $(L) elements.")
+operator_eltype(::FilterTypeOperator{R}, ::Type{R}) where { R }         = R
+operator_eltype(::FilterTypeOperator{R}, ::Type{L}) where { L, R <: L } = R
+operator_eltype(::FilterTypeOperator{R}, ::Type{L}) where { L, R }      = error("filter_type(T) operator creation error. Cannot filter $(R) elements from stream of $(L) elements.")
 
-struct FilterTypeProxy{L, T} <: ActorProxy end
+struct FilterTypeSubscribable{R, S} <: Subscribable{R} 
+    source :: S
+end
 
-actor_proxy!(::Type{T}, proxy::FilterTypeProxy{L, T}, actor::A) where { L, A, T } = FilterTypeActor{L, A, T}(actor)
+function on_subscribe!(subscribable::FilterTypeSubscribable{R}, actor::A) where { R, A }
+    return subscribe!(subscribable.source, FilterTypeActor{R, A}(actor))
+end
 
-struct FilterTypeActor{L, A, T} <: Actor{L}
+struct FilterTypeActor{R, A}
     actor :: A
 end
 
-on_next!(actor::FilterTypeActor{L, A, T}, data::T) where { L, A, T } = next!(actor.actor, data)
-on_next!(::FilterTypeActor, data)                                    = nothing
+on_next!(actor::FilterTypeActor{R, A}, data::R) where { R, A } = on_next!(actor.actor, data)
+on_next!(::FilterTypeActor, data)                              = begin end
 
-on_error!(actor::FilterTypeActor, err) = error!(actor.actor, err)
-on_complete!(actor::FilterTypeActor)   = complete!(actor.actor)
+on_error!(actor::FilterTypeActor, err) = on_error!(actor.actor, err)
+on_complete!(actor::FilterTypeActor)   = on_complete!(actor.actor)
 
-Base.show(io::IO, ::FilterTypeOperator)         = print(io, "FilterTypeOperator()")
-Base.show(io::IO, ::FilterTypeProxy)            = print(io, "FilterTypeProxy()")
-Base.show(io::IO, ::FilterTypeActor{L}) where L = print(io, "FilterTypeActor($L)")
+Base.show(io::IO, ::FilterTypeOperator)                = print(io, "FilterTypeOperator()")
+Base.show(io::IO, ::FilterTypeSubscribable{R}) where R = print(io, "FilterTypeSubscribable($R)")
+Base.show(io::IO, ::FilterTypeActor{R}) where R        = print(io, "FilterTypeActor($R)")