+++ src/operators/map.jl
@@ -2,7 +2,7 @@
 import Base: map, show
 
 """
-    map(::OpType{R}, mappingFn::F) where { F }
+    map(::OpType{R}, mapping::F) where { F }
 
 Creates a map operator, which applies a given `mappingFn` callback to each value emmited by the source
 Observable, and emits the resulting values as an Observable. You have to specify output `R` type after
@@ -10,7 +10,7 @@ Observable, and emits the resulting values as an Observable. You have to specify
 
 # Arguments
 - `::OpType{R}`: the type of data of transformed value, may be or may not be the same as source type
-- `mappingFn::Function`: transformation function with `(data::L) -> R` signature, where L is type of data in input source
+- `mapping::F`: transformation callable object with `(data::L) -> R` signature, where L is type of data in input source
 
 # Producing
 
@@ -32,7 +32,7 @@ subscribe!(source |> map(OpType(Int), (d) -> d ^ 2), logger())
 [LogActor] Completed
 ```
 
-See also: [`AbstractOperator`](@ref), [`RightTypedOperator`](@ref), [`ProxyObservable`](@ref), [`logger`](@ref)
+See also: [`Operator`](@ref), [`Subscribable`](@ref), [`OpType`](@ref)
 """
 map(::OpType{R}, mapping::F) where { R, F } = MapOperator{R, F}(mapping)
 
@@ -45,7 +45,7 @@ struct MapSubscribable{R, F, S} <: Subscribable{R}
     source  :: S
 end
 
-struct MapActor{R, F, A} <: Actor{Any}
+struct MapActor{R, F, A}
     mapping :: F
     actor   :: A
 end
@@ -54,13 +54,13 @@ function on_call!(_, ::Type{R}, operator::MapOperator{R, F}, source::S) where { 
     return MapSubscribable{R, F, S}(operator.mapping, source)
 end
 
-function on_subscribe!(source::MapSubscribable{R, F}, actor::A, scheduler) where { R, F, A }
-    return on_subscribe!(source.source, MapActor{R, F, A}(source.mapping, actor), scheduler)
+function on_subscribe!(source::MapSubscribable{R, F}, actor::A) where { R, F, A }
+    return subscribe!(source.source, MapActor{R, F, A}(source.mapping, actor))
 end
 
-next!(actor::MapActor{R}, data) where R = next!(actor.actor, convert(R, actor.mapping(data)))
-error!(actor::MapActor, err)            = error!(actor.actor, err)
-complete!(actor::MapActor)              = complete!(actor.actor)
+on_next!(actor::MapActor{R}, data) where R = on_next!(actor.actor, convert(R, actor.mapping(data)))
+on_error!(actor::MapActor, err)            = on_error!(actor.actor, err)
+on_complete!(actor::MapActor)              = on_complete!(actor.actor)
 
 Base.show(io::IO, ::MapOperator{R})     where R   = print(io, "MapOperator( -> $R)")
 Base.show(io::IO, ::MapSubscribable{R}) where R   = print(io, "MapSubscribable( -> $R)")