+++ src/schedulers/asap.jl
@@ -13,14 +13,11 @@ See also: [`AsapSchedulerInstance`](@ref)
 struct AsapScheduler <: Scheduler end
 
 Base.show(io::IO, ::AsapScheduler) = print(io, "AsapScheduler()")
+Base.similar(::AsapScheduler)      = AsapScheduler()
 
-Base.similar(::AsapScheduler) = AsapScheduler()
+subscribe!(::AsapScheduler, subscribable, actor) = on_subscribe!(subscribable, actor)
+unsubscribe!(::AsapScheduler, subscription)      = on_unsubscribe!(subscription)
 
-struct AsapSchedulerInstance end
-
-makeinstance(_, ::AsapScheduler)         = AsapSchedulerInstance()
-instancetype(_, ::Type{ AsapScheduler }) = AsapSchedulerInstance
-
-next!(actor, value, ::AsapSchedulerInstance) = next!(actor, value)
-error!(actor, err, ::AsapSchedulerInstance)  = error!(actor, err)
-complete!(actor, ::AsapSchedulerInstance)    = complete!(actor)
+next!(::AsapScheduler, actor, value) = on_next!(actor, value)
+error!(::AsapScheduler, actor, err)  = on_error!(actor, err)
+complete!(::AsapScheduler, actor)    = on_complete!(actor)